chore: resolve merge conflict in generate route
